+++ packages/neuron-wallet/src/models/subjects/command.ts
@@ -2,8 +2,9 @@ import { Subject } from 'rxjs'
 
 const CommandSubject = new Subject<{
   winID: number
-  type: 'nav' | 'delete-wallet' | 'backup-wallet'
+  type: 'navigate-to-url' | 'delete-wallet' | 'backup-wallet' | 'export-xpubkey' | 'import-xpubkey'
   payload: string | null
+  dispatchToUI: boolean
 }>()
 
 export default CommandSubject